% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExample_basic}
\alias{guideSetExample_basic}
\title{Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
the CDS of the human gene KRAS}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(guideSetExample_basic, package="crisprShiny")
}
\description{
Example of a \linkS4class{GuideSet} object (with no additional annotation)
storing gRNA sequences targeting the coding sequence of human gene KRAS
(ENSG00000133703) for SpCas9 nuclease.
}
\details{
The object was obtained by using \code{crisprDesign::findSpacers}
on a GRanges of the CDS region of human gene KRAS. See code in
\code{inst/scripts/generateKrasData.R}.
}
\keyword{datasets}
