% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExample_kras_be}
\alias{guideSetExample_kras_be}
\title{Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
the CDS of the human gene KRAS}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(guideSetExample_kras_be, package="crisprShiny")
}
\description{
Example of a \linkS4class{GuideSet} object storing gRNA
sequences targeting the coding sequence of human gene KRAS
(ENSG00000133703) for BE4max nuclease.
}
\details{
The object was obtained by applying all base-editor-specific
annotation functions in \code{crisprDesign} on the
CDS region of human gene KRAS. See code in
\code{inst/scripts/generateKrasData.R}.
}
\keyword{datasets}
