% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExample_ntcs}
\alias{guideSetExample_ntcs}
\title{Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
the CDS of the human gene KRAS and NTCs}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(guideSetExample_kras, package="crisprShiny")
}
\description{
Example of a fully annotated \linkS4class{GuideSet} object storing gRNA
sequences targeting the coding sequence of human gene KRAS
(ENSG00000133703) and some non-targeting controls (NTCs) for SpCas9
nuclease.
}
\details{
The object was obtained by applying all available \code{add*}
annotation functions in \code{crisprDesign} on a GuideSet storing gRNAs
targeting the CDS region of human gene KRAS and some NTCs. See code in
\code{inst/scripts/generateKrasData.R}.
}
\keyword{datasets}
