% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tooltipAnnotation}
\alias{tooltipAnnotation}
\title{List of tooltip annotations}
\format{
A \code{list} object.
}
\usage{
data(tooltipAnnotation, package="crisprShiny")
}
\description{
Tooltip messages for crisprShiny app stored in a \code{list} object.
}
\details{
See code in \code{inst/scripts/generateKrasData.R} for tooltip
annotations.
}
\keyword{datasets}
