% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tss_kras}
\alias{tss_kras}
\title{Example of a \code{GenomicRanges} object storing annotated TSS ranges
for the human gene KRAS}
\format{
A \code{GenomicRanges} object.
}
\usage{
data(tss_kras, package="crisprShiny")
}
\description{
Example of a \code{GenomicRanges} object storing annotated TSS ranges
for the human gene KRAS (ENSG00000133703).
}
\details{
The object was obtained by subsetting genomic ranges in
\code{tss_human} from the \code{crisprDesignData} package for the
KRAS gene (Ensembl ID: ENSG00000133703). See code in
\code{inst/scripts/generateKrasData.R}.
}
\keyword{datasets}
