% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb_kras}
\alias{txdb_kras}
\title{Example of a \code{CompressedGenomicRangesList} object storing
annotated ranges for the human gene KRAS}
\format{
A \code{CompressedGenomicRangesList} object.
}
\usage{
data(txdb_kras, package="crisprShiny")
}
\description{
Example of a \code{CompressedGenomicRangesList} object storing
annotated genomic ranges for the human gene KRAS (ENSG00000133703).
}
\details{
The object was obtained by subsetting all genomic ranges annotations
in \code{txdb_human} from the \code{crisprDesignData} package for the
KRAS gene (Ensembl ID: ENSG00000133703). See code in
\code{inst/scripts/generateKrasData.R}.
}
\keyword{datasets}
